backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
SQLAlchemy>=2.0.32
psycopg[binary]>=3.2
openai>=1.0.0
supabase>=2.4.0
python-jose[cryptography]>=3.3.0
